requirements.txt
hf_xet
ffmpeg-python
faster-whisper
